qcacld-3.0: Add BW check for whitelist APs to disable DSMPS
For STA connection with DSMPS Whitelisted APs on 160 MHz, AP sends
RTS on 160 MHz instead of 20 MHz and with low RSSI, STA misses this
frame and CTS is not sent hence plummeting the device throughput for
lower RSSI levels. Hence disable DSMPS in HT and HE capability IEs
in association request for 160 MHz bandwidth for whitelisted APs.

Change-Id: Id64221612bf0c25da68910db704609ca0533c8ba
CRs-Fixed: 4313114