qcacld-3.0: Configure TDLS offchannel based on concurrency
In non-dbs solutions, the TDLS off-channel is not disabled
when STA and SAP interfaces are running on SCC. This puts
the firmware in an undesired state. SAP + SAP + TDLS is
allowed only without off-channel.

To fix this, disable the off-channel if STA+SAP is running
in SCC on a non-dbs solution. Enable the off-channel back,
if the SAP interface is down.

Change-Id: If768bb63820f6c85cd16501e46792948d616ea7c
CRs-Fixed: 4313422