qcacld-3.0: Skip SAP+SAP scc check for GO + LT SAP
When initiating a GO start in non-DBS, with XPAN present,
the checks for SAP and GO SCC kicks in, allowing SCC to proceed.

Thus skip this step and continue to select MCC.

Additionally, to prevent CAC during LL LT SAP's operation,
avoid allowing GO in the 160 MHz band if XPAN is present.

CRs-Fixed: 4315659
Change-Id: I9c64eb0c9373fe7bc4fb0194b0543e810ec3af45