qcacld-3.0: Define band bitmap with qdf_bitmap
Currently, band bitmap is defined as an unsigned long, kernel
API test_bit is used to manipulate it. However, these API might
expect an unsigned long array, and a static analyzer tool has
flagged this as an out-of-bounds (OOB) access.

To address this, use qdf_bitmap to define and operate on the
bitmap, ensuring proper handling.

Change-Id: I4c8fb06c2eea59b94e64e399274dec250d62a05c
CRs-Fixed: 4321719