qcacld-3.0: Fix NULL pointer dereference on peer stats command
While roaming is in progress, a station dump request was received.
Due to the MAC address change during roaming, the vdev became NULL
when attempting to send the peer stats command to the firmware,
resulting in a NULL pointer dereference.

To prevent this, a NULL check is added before sending the peer stats
request to the firmware.

Change-Id: Id3386ff40ff434046d6d43512ae0f032eb8e36dd
CRs-Fixed: 4316511