qcacld-3.0: Don't enable roaming when link switch is in progress
When ML STA + P2P CLI concurrency exists, if the P2P CLI goes
down, then the roaming will be enabled on the concurrent
STA vdev. But if link switch is in progress at the same time,
if P2P vdev peer delete is done first, then RSO start goes
for the STA vdev. This causes the Roam scan mode command to be
sent with mode = 7 before vdev stop is sent to firmware for the
link switch command. This causes firmware to enable roam scans
while link switch is in progress and leads to abnormal behavior.

Reset the requestor bitmap and don't send RSO start if link
switch is in progress.

CRs-Fixed: 4287287
Change-Id: Iae987fa2e864af0d5dc874b9798204e83c0f4005