qcacld-3.0: Intersect Host and Firmware CCK Support
This update introduces a mechanism to conditionally send
the VDEV parameter for CCK 5 GHz support to the firmware.
The decision to send this parameter is now based on the
intersection of Host INI configuration and Firmware CCK
capability. This ensures that the feature is only enabled
when both Host and Firmware support it, thereby improving
compatibility and reducing misconfigurations.
The enhancement includes:

a.)Parsing FW capability.
b.) Using bitmask operations to extract relevant CCK
support bits for all mode.
c.) Updating the INI to include CCK support parameters
based on intersection.

Change-Id: Ib558c43ed7a7f393e6fcec709b52d549f10d7971
CRs-Fixed: 4329030