qcacld-3.0: Skip PMKID assignment in WPA2 sessions
Currently, the association request frame for WPA2 connections
includes a cached or previous PMKID. This causes the AP to reject
the connection with an unspecified failure status code, leading to
connection issues.

To resolve this, skip sending the PMKID in WPA2 association
requests, especially when PMKSA is zero. This ensures compatibility
with APs that expect no PMKID in WPA2 scenarios and prevents
unnecessary connection failures.

Change-Id: I11341ed3cf450f50e103c1a7f07c1d5289df2847
CRs-Fixed: 4338996