qcacld-3.0: Set install key flag when peer is created
Issue is DUT transmitted a DSCP policy response frame without
enabling Protected Flag. This scenario occurred when DUT is
roamed to a different BSSID. When roam sync callback is called
it tries to set the flag to true but it cannot be done since the
peer object is not created at that moment.
This code is to set key installed flag when new peer object is created.

CRs-Fixed: 4274743
Change-Id: If95920b92488656823d727a757cf181b5cf8c833