qcacld-3.0: Limit MCC GO to 80 MHz in 5 GHz
When HE is enabled for P2P, the supplicant initiates the GO on
a 160 MHz channel in the 5 GHz band. Consequently, starting the
GO on a 160 MHz channel in the 5 GHz band can lead to MCC if any
interface is present, resulting in the rejection of the GO due
to the inability to support DFS beaconing in MCC.

To resolve this, downgrade the GO to 80 MHz in such cases.

Change-Id: Ia7057637609efcf85e4cad4eaec14dfe0e3d1967
CRs-Fixed: 4558456