qcacld-3.0: Fix invalid 11mc responder bits conversion
To enable 11mc in firmware, the gfine_time_meas_cap ini is read
and based on the SAP responder bits the converted bitmap which
should be sent to firmware is filled. But there is a wrong
conversion logic while checking for the 11az TB & NTB ranging
bits. This results in enabling 11mc even when the
gfine_time_meas_cap bitmap has SAP responder role disabled

Fix the wrong conversion associated with the filling of
responder bits.

CRs-Fixed: 4332458
Change-Id: I16cd29b70acc3280aa92483f3f6ef8860baaaf80