qcacld-3.0: Remove bus bw votes when bus bw timer is stopped
When the bus bandwidth timer is stopped, and when there are no
active connections, the bus bandwidth votes are removed to save
power. This logic is broken because of the DP componentization
change and the votes are applied even when no active connections
are present.

When the bus bandwidth timer is stopped, irrespective of the
connection status, remove the votes since voting in suspend
state is not needed.

Change-Id: I540ee70ad83cacdb693340cc90360ad280cd7241
CRs-Fixed: 4361528