qcacld-3.0: Fix NAN ndi peer migration race condition
Currently, when we bring down the wifi-aware0 interface using
ifconfig, the NDI peer and the vdev are deleted in different
contexts. Because of this, during NDP cleanup we delete the
NDI peer that has been migrated, and at the same time, when
deleting the vdev, we check if there are any PASN peers.

Since we have already deleted the NDI peer and have not yet
recreated the PASN peer, we proceed with vdev deletion. However,
in a different context, the PASN peer is then recreated and
holds a reference to the vdev object. As a result, the vdev is
deleted logically, but cannot be deleted physically because the
reference is still held.

Fix this by waiting for the peer migration to complete before
deleting the vdev.

Change-Id: I5e5f9d3a3b0db53721c42eba60a1885cad439e14
CRs-Fixed: 4374920