qcacld-3.0: Move SAP out of MCC even if it's on unsafe channel
Currently, fixed channel SAP is allowed to stay on the
operating channel if it's marked as unsafe channel. This causes
SAP to stay in MCC with STA if STA moves from an unsafe channel
to a non-allowed channel(e.g. DFS, 6 GHz, indoor).
As "MCC-to-SCC switch" takes priority compared to "fixed channel
SAP preference on unsafe channel", allow SAP to move out of MCC.

Change-Id: Ib506f1e283ff64a82a96eb599df9436fbc136d77
CRs-Fixed: 4381970