qcacld-3.0: Move SAP to original channel even if it's an unsafe ch
Currently, SAP is moved to 2 GHz when NAN/NDP concurrency come up
for the best utilization of 5 GHz band for NDP throughput cases.
SAP is moved back to its original channel(e.g. 5 GHz channel) when
NAN usecase is done, i.e. when NAN is disabled.
But SAP doesn't move to the original channel if it's marked as
unsafe(due to LTE coex operations) channel.
Allow SAP to move to original channel event if it's unsafe when
required conditions are met(i.e. it's a fixed channel SAP and
FW capability is also present)

Change-Id: I2f3d38861de52618e02522a891e5ff3b29ec62c3
CRs-Fixed: 4384341