qcacld-3.0: Ignore get_sta for previous AP during roam
The get_station command is received for the old bssid
when the host driver has already completed roaming to
a new AP. However, the roamed event is not yet updated
to the supplicant. Since, the VDEV SM is also UP after
roaming, the get_station request is processed and the
WMI_REQUEST_PEER_STATS_INFO_CMDID is sent to the firmware
for the old BSSID. FW does not respond to the STATS cmd
leading to timeout after 1.5seconds. During this time, the
wpa_supplicant thread gets suspended and there is a delay
in processing the roam event and EAPOL frames leading to
disconnection.

To fix this, add a check to validate the bssid for the
legacy case, similar to the existing check for MLO links.

Change-Id: Id1509ce6d954dff5b29a275d0757de0af6af9fff
CRs-Fixed: 4381772